Testing/polybot/requirements.txt
py-clob-client
python-dotenv
eth-account
websockets
requests